Abstract place_ships to Player class
