+++ lib/battleship_game.rb
@@ -38,6 +38,13 @@ class BattleshipGame
     @start_time = Time.now
 
     @player_2.place_ships
+
+    puts %{I have laid out my ships on the grid.
+You now need to layout your two ships.
+The first is two units long and the
+second is three units long.
+The grid has A1 at the top left and D4 at the bottom right.}
+
     @player_1.place_ships
 
     puts "May the best player win!"
